add hardware encoding

+++ install.sh
@@ -13,54 +13,103 @@ fi
 source ./config
 
 
-install=true
+sysconf=false
 update=false
+dlrepo=false
+buildnvcodec=false
+makemkv=false
+ffmpeg=false
 
-while getopts ":uif:m:" flag
+Help()
+{
+   # Display Help
+   echo "Add description of the script functions here."
+   echo
+   echo "options:"
+   echo "h              Print this Help."
+   echo "s              configure systemfiles and apt packages."
+   echo "b              install ripping scripts."
+   echo "d              download updated ripping scripts from repo."
+   echo "f  opt<ver>    build and install ffmpeg.  optional version number."
+   echo "m  opt<ver>    build and install makemkvcon. option version number."
+   echo "n              build nv-codec for gpu acceleration."
+   echo
+}
+
+getopts_get_optional_argument() {
+  eval next_token=\${$OPTIND}
+  if [[ -n $next_token && $next_token != -* ]]; then
+    OPTIND=$((OPTIND + 1))
+    OPTARG=$next_token
+  else
+    OPTARG=""
+  fi
+}
+
+while getopts ":hsbdfmn" flag
 do
-    case ${flag} in
-        u) update=true
-           install=false
-           ;;
-        i) install=true
-           update=false
-           ;;
-        f) ffmpegVer=${OPTARG}
-           ;;
-        m) makemkvVer=${OPTARG}
-           ;;
-        *)
-                echo 'Error in command line parsing' >&2
-                exit 1
-    esac
+  case ${flag} in
+      
+    h) # display Help
+       Help
+       exit;;
+    s) sysconf=true;;
+    b) bins=true;;
+    d) dlrepo=true;;
+    f) ffmpeg=true
+       getopts_get_optional_argument $@
+       ffmpegVer=${OPTARG};;
+    
+    m) makemkv=true
+       getopts_get_optional_argument $@ 
+       makemkvVer=${OPTARG};;
+    n) buildnvcodec=true ;;
+    *)
+        echo 'Error in command line parsing' >&2
+        exit 1
+  esac
 done
 
+if [ "$opt" = "?" ]
+then
+  echo "Default option executed (by default)"
+  
+  sysconf=true
+  update=true
+  dlrepo=true
+  buildnvcodec=true
+  makemkv=true
+  ffmpeg=true
+fi
 
 homedir="/home/${user}"
 
-if [[ $install == true ]]; then
- echo "install"
+if [[ $sysconf == true ]]; then
+ echo "config system"
   apt-get install libdvd-pkg
   dpkg-reconfigure libdvd-pkg
-  apt-get install regionset libavcodec-extra dvdbackup yasm lsdvd autofs abcde at flac mkvtoolnix
+  apt-get install regionset libavcodec-extra dvdbackup yasm lsdvd autofs abcde at flac git
   apt-get install build-essential pkg-config libc6-dev libssl-dev libexpat1-dev libavcodec-dev libgl1-mesa-dev qtbase5-dev zlib1g-dev libx264-dev libx265-dev
 
   systemctl enable --now atd
   
-
   echo 'SUBSYSTEM=="block", ENV{ID_CDROM}=="1", ENV{ID_TYPE}=="cd", ACTION=="change", RUN+="/usr/local/bin/autodisk '"'"'%E(DEVNAME)'"'"'"' | tee /etc/udev/rules.d/autodisk.rules
 
   udevadm control --reload
 
 fi
 
+if [[ $bins == true ]]; then
 
-if [[ $install == true ]] || [[ $update == true ]]; then
-
-  wget -O ./cdrip https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/cdrip
-  wget -O ./dvdrip https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/dvdrip
-  wget -O ./bdrip https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/bdrip
-  wget -O ./autodisk https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/autodisk
+  echo "install binaries"
+  
+  if [[ $dlrepo == true ]]; then
+    echo "download binaries from repo"
+    wget -O ./cdrip https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/cdrip
+    wget -O ./dvdrip https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/dvdrip
+    wget -O ./bdrip https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/bdrip
+    wget -O ./autodisk https://raw.githubusercontent.com/nathanjshaffer/diskripper/main/autodisk
+  fi  
   
   echo -e '#!/bin/bash'"\n" > /tmp/cdrip
   echo "user=\"${user}\"
@@ -88,34 +137,47 @@ outformat=\"${BD_outformat}\"" >> /tmp/bdrip
   echo "user=\"${user}\"" >> /tmp/autodisk
   cat ./autodisk >> /tmp/autodisk
 
-
   install -c -D -m 755 /tmp/autodisk /usr/local/bin/
   install -c -D -m 755 /tmp/dvdrip /usr/local/bin/
   install -c -D -m 755 /tmp/cdrip /usr/local/bin/
   install -c -D -m 755 /tmp/bdrip /usr/local/bin/
-  
-  
-  if ! ffmpeg -version | grep -q "ffmpeg version ${ffmpegVer}";
+fi
+
+if [[ $buildnvcodec == true ]]; then
+  echo "build nv-codec"
+  pdir="$PWD"
+  mkdir ./nv-codec-headers_build && cd ./nv-codec-headers_build
+  git clone https://git.videolan.org/git/ffmpeg/nv-codec-headers.git
+  cd nv-codec-headers
+  make && sudo make install
+  cd "$pdir"
+fi
+
+
+if  ! ffmpeg -version | grep -q "ffmpeg version ${ffmpegVer}" && [[ $ffmpeg == true ]] ;
   then
     echo "downloading ffmpeg"
+    
+    pdir="$PWD"
     wget https://ffmpeg.org/releases/ffmpeg-${ffmpegVer}.tar.xz
     tar -xvf ffmpeg-${ffmpegVer}.tar.xz
     
     cd ffmpeg-${ffmpegVer}
-    echo "$PWD"
     
-    ./configure --enable-gpl --enable-libx264 --enable-libx265
+    ./configure --enable-gpl --enable-libx264 --enable-libx265 --enable-nvenc
     make
     make install
     rm -rf /tmp/ffmpeg
 
-    cd../
+    cd "$pdir"
   else
     echo "ffmpeg is current version. Skipping..."
-  fi
+fi
 
-  if [ ! -d "./makemkv-bin-${makemkvVer}" ];
+if [ ! -d "./makemkv-bin-${makemkvVer}" ] && [[ $makemkv == true ]];
   then
+    echo "downloading makemkv"
+    pdir="$PWD"
     wget https://www.makemkv.com/download/makemkv-bin-${makemkvVer}.tar.gz
     wget https://www.makemkv.com/download/makemkv-oss-${makemkvVer}.tar.gz
 
@@ -131,10 +193,9 @@ outformat=\"${BD_outformat}\"" >> /tmp/bdrip
     make
     make install
 
-    cd ../
+    cd "$pdir"
   else
     echo "makemkv is current version. Skipping..."
-  fi
 fi
